excel-auto/requirements.txt
pandas
openpyxl
matplotlib
pyyaml